Fix warnings (v5.47.1)
- Drop unnecessary parentheses in aifw-ids/src/protocol/dns.rs that
  rustc was flagging under unused_parens.
- Apply cargo clippy --fix across the workspace: collapse nested
  `if let`s, drop empty format-string literals, remove redundant
  closures, fix manual prefix stripping, and other stylistic nits.
  No behavioural changes.

`cargo build --workspace` and `cargo test --workspace --no-run`
both run with zero warnings.

+++ aifw-daemon/src/main.rs
@@ -147,7 +147,7 @@ async fn main() -> anyhow::Result<()> {
                     let sample = aifw_core::multiwan::SlaSample {
                         gateway_id: g.id,
                         bucket_ts: bucket,
-                        samples: 60 / (g.interval_ms.max(1) as u64 / 1000).max(1),
+                        samples: 60 / (g.interval_ms.max(1) / 1000).max(1),
                         rtt_avg: g.last_rtt_ms,
                         rtt_p95: g.last_rtt_ms,
                         rtt_p99: g.last_rtt_ms,
